+++ src/spaemis/commands/project_command.py
@@ -11,14 +11,18 @@ from spaemis.scaling import get_scaler_by_config
 logger = logging.getLogger(__name__)
 
 
-def scale_inventory(cfg: VariableConfig, inventory: EmissionsInventory):
+def scale_inventory(
+    cfg: VariableConfig, inventory: EmissionsInventory, target_year: int
+):
     if cfg.variable not in inventory.data.variables:
         raise ValueError(f"Variable {cfg.variable} not available in inventory")
     if cfg.sector not in inventory.data["sector"]:
         raise ValueError(f"Sector {cfg.sector} not available in inventory")
     field = inventory.data[cfg.variable].sel(sector=cfg.sector)
 
-    scaled_field = get_scaler_by_config(cfg.method)(field)
+    scaled_field = get_scaler_by_config(cfg.method)(
+        field, inventory=inventory, target_year=target_year
+    )
 
     return scaled_field
 
@@ -34,7 +38,6 @@ def run_project_command(config, out_dir):
     """
     Generate a set of emissions projection using an emissions inventory as a base
     """
-
     config = load_config(config)
 
     inventory = load_inventory(config.inventory_name, year=config.inventory_year)
@@ -46,4 +49,4 @@ def run_project_command(config, out_dir):
     for slice_year in config.timeslices:
         logger.info(f"Processing year={slice_year}")
         for projection_config in config.variables:
-            scale_inventory(projection_config, inventory)
+            scale_inventory(projection_config, inventory, slice_year)