+++ CHANGELOG.md
@@ -17,5 +17,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add relative_change scaler and reading of Input4MIPs data [#3](https://github.com/climate-resource/spaemis/pull/3)
 - Added CLI command `project` and a framework for scalers [#2](https://github.com/climate-resource/spaemis/pull/2)
 - Initial commit and repository setup